Internals

Internals

LogDensityProblems.AbstractLogDensityProblem — Type.

Abstract type for log density representations, which support the following interface for ℓ::AbstractLogDensityProblem:

dimension returns the dimension of the domain of ℓ,

logdensity evaluates the log density ℓ at a given point.

See also LogDensityProblems.stresstest for stress testing.

source

LogDensityProblems.LogDensityWrapper — Type.

An abstract type that wraps another log density in its field ℓ.

Notes

Implementation detail, not exported.

source

LogDensityProblems.ADGradientWrapper — Type.

An abstract type that wraps another log density for calculating the gradient via AD.

Automatically defines a logdensity(Value, ...) method, subtypes should define a logdensity(ValueGradient, ...) one.

source

LogDensityProblems.heuristic_chunks — Function.

heuristic_chunks(N)
heuristic_chunks(N, M)

Default chunk sizes to try for benchmarking. Fewer than M, always contains 1 and N.

source

LogDensityProblems.RejectLogDensity — Type.

struct RejectLogDensity <: Exception

Exception for unwinding the stack early for infeasible values. Use reject_logdensity().

source