Home

Introduction

This package provides the following functionality:

It defines the logdensity method with corresponding interface, which can be used by other packages that operate on (log) densities and need to evaluate the log densities or the gradients (eg MCMC, MAP, ML or similar methods).

It defines the ADgradient which makes objects that support logdensity to calculate log density values calculate log density gradients using various automatic differentiation packages.

It defines the wrapper TransformedLogDensity using the TransformVariables.jl package, allowing callables that take a set of parameters transformed from a flat vector of real numbers to support the logdensity interface.

Various utility functions for debugging and testing log densities.

Inference

LogDensityProblems.logdensity — Function.

logdensity(resulttype, ℓ, x)

Evaluate the AbstractLogDensityProblem ℓ at x, which has length compatible with its dimension.

The argument resulttype determines the type of the result:

Real for an unchecked evaluation of the log density which should return a ::Real

number (that could be NaN, Inf, etc),

Value for a checked log density, returning a Value,

ValueGradient also calculates the gradient, returning a ValueGradient,

ValueGradientBuffer calculates a ValueGradient potentially (but always

consistently for argument types) using the provided buffer for the gradient. In this case, the element type of the array may determine the result element type.

Implementation note

Most types should just define the methods for Real and ValueGradientBuffer (when applicable), as Value and ValueGradient fall back to these, respectively.

source

TransformVariables.dimension — Function.

dimension(p)

The dimension of the problem, ie the length of the vectors in its domain.

source

LogDensityProblems.Value — Type.

Value(value)

Holds the value of a logdensity at a given point.

Constructor ensures that the value is either finite, or $-∞$.

All other values (eg NaN or Inf for the value) lead to an error.

See also logdensity.

source

LogDensityProblems.ValueGradient — Type.

ValueGradient(value, gradient)

Holds the value and gradient of a logdensity at a given point.

Constructor ensures that either

both the value and the gradient are finite,

the value is $-∞$ (then gradient is not checked).

All other values (eg NaN or Inf for the value) lead to an error.

See also logdensity.

source

LogDensityProblems.ValueGradientBuffer — Type.

struct ValueGradientBuffer{T<:Real, V<:AbstractArray{T<:Real,1}}

A wrapper for a vector that indicates that the vector may be used for the gradient in a ValueGradient. Consequences are undefined if it is modified later, implicitly the caller guarantees that it will not be used for anything else while the gradient is retrieved.

See logdensity.

source

Gradient via automatic differentiation

LogDensityProblems.ADgradient — Function.

ADgradient(kind, P; kwargs...)

Wrap P using automatic differentiation to obtain a gradient.

kind is usually a Val type, containing a symbol that refers to a package. The symbol can also be used directly as eg

ADgradient(:ForwardDiff, P)

See methods(ADgradient). Note that some methods are defined conditionally on the relevant package being loaded.

source

ADgradient(#temp#, ℓ; chunk, gradientconfig)

Wrap a log density that supports evaluation of Value to handle ValueGradient, using ForwardDiff.

Keywords are passed on to ForwardDiff.GradientConfig to customize the setup. In particular, chunk size can be set with a chunk keyword argument (accepting an integer or a ForwardDiff.Chunk).

source

ADgradient(?, ℓ)

Gradient using algorithmic/automatic differentiation via Flux.

source

Transformed problem definition

LogDensityProblems.TransformedLogDensity — Type.

TransformedLogDensity(transformation, log_density_function)

A problem in Bayesian inference. Vectors of length dimension(transformation) are transformed into a general object θ (unrestricted type, but a named tuple is recommended for clean code), correcting for the log Jacobian determinant of the transformation.

It is recommended that log_density_function is a callable object that also encapsulates the data for the problem.

log_density_function(θ) is expected to return real numbers. For zero densities or infeasible θs, -Inf or similar should be returned, but for efficiency of inference most methods recommend using transformation to avoid this.

Use the property accessors ℓ.transformation and ℓ.log_density_function to access the arguments of ℓ::TransformedLogDensity, these are part of the API.

source

Benchmarking, diagnostics, and utilities

LogDensityProblems.stresstest — Function.

stresstest(ℓ; N, rng, scale, resulttype)

Test ℓ with random values.

N random vectors are drawn from a standard multivariate Cauchy distribution, scaled with scale (which can be a scalar or a conformable vector). In case the call produces an error, the value is recorded as a failure, which are returned by the function.

Not exported, but part of the API.

source

LogDensityProblems.benchmark_ForwardDiff_chunks — Function.

benchmark_ForwardDiff_chunks(ℓ; chunks, resulttype, markprogress)

Benchmark a log density problem with various chunk sizes using ForwardDiff.

chunks, which defaults to all possible chunk sizes, determines the chunks that are tried.

The function returns chunk => time pairs, where time is the benchmarked runtime in seconds, as determined by BenchmarkTools.@belapsed.

Runtime may be long because of tuned benchmarks, so when markprogress == true (the default), dots are printed to mark progress.

This function is not exported, but part of the API.

source

LogDensityProblems.@iffinite — Macro.

If expr evaluates to a non-finite value, return with that, otherwise evaluate to that value. Useful for returning early from non-finite likelihoods.

Part of the API, but not exported.

source

LogDensityProblems.LogDensityRejectErrors — Type.

LogDensityRejectErrors(ℓ)

Wrap a logdensity ℓ so that errors <: E are caught and replaced with a $-∞$ value.

E defaults to InvalidLogDensityExceptions.

Note

Use cautiously, as catching errors can mask errors in your code. The recommended use case is for catching quirks and corner cases of AD. See also stresstest as an alternative to using this wrapper.

Also, some AD packages don't handle try - catch blocks, so it is advised to use this as the outermost wrapper.

source